+++ src/model/PomodoroTimer.ts
@@ -1,4 +1,4 @@
-import { convertToKoreaDate } from "../util/dateUtil.js";
+import { convertToKoreaDate } from "../util/date_util.js";
 
 // TODO: 클라이언트랑 같이 쓰이는 클래스임 때문에 한군데다 모으고 동시에 쓰는 방법을 찾아야함
 export enum PomodoroTimerState {
