+++ tsconfig.json
@@ -14,7 +14,7 @@
   "include": [
     "src/**/*.ts",
     "src/model/PomodoroTimer.ts",
-    "src/util/dateUtil.ts"
+    "src/util/date_util.ts"
   ],
   "exclude": [
     "node_modules"
